Tuck node signal handlers behind build tags
